Retreiving actors for group members
